@@ -38,15 +38,16 @@ use futures::{future, stream, Sink, SinkExt, Stream, StreamExt, TryFutureExt};
 use number_prefix::NumberPrefix;
 use serde::{Deserialize, Serialize};
 use std::cell::Cell;
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 use std::env;
-use std::ffi::{OsStr, OsString};
+use std::ffi::OsString;
 use std::future::Future;
 use std::io::{self, Write};
 use std::marker::Unpin;
 #[cfg(feature = "dist-client")]
 use std::mem;
-use std::net::{Ipv4Addr, SocketAddr, SocketAddrV4};
+#[cfg(any(target_os = "linux", target_os = "android"))]
+use std::os::linux::net::SocketAddrExt;
 use std::path::PathBuf;
 use std::pin::Pin;
 use std::process::{ExitStatus, Output};
@@ -55,12 +56,10 @@ use std::task::{Context, Poll, Waker};
 use std::time::Duration;
 #[cfg(feature = "dist-client")]
 use std::time::Instant;
-use std::u64;
 use tokio::sync::Mutex;
 use tokio::sync::RwLock;
 use tokio::{
     io::{AsyncRead, AsyncWrite},
-    net::TcpListener,
     runtime::Runtime,
     time::{self, sleep, Sleep},
 };
@@ -81,8 +80,8 @@ const DIST_CLIENT_RECREATE_TIMEOUT: Duration = Duration::from_secs(30);
 /// Result of background server startup.
 #[derive(Debug, Serialize, Deserialize)]
 pub enum ServerStartup {
-    /// Server started successfully on `port`.
-    Ok { port: u16 },
+    /// Server started successfully on `addr`.
+    Ok { addr: String },
     /// Server Addr already in suse
     AddrInUse,
     /// Timed out waiting for server startup.
@@ -145,7 +144,7 @@ pub struct DistClientContainer {
 }
 
 #[cfg(feature = "dist-client")]
-struct DistClientConfig {
+pub struct DistClientConfig {
     // Reusable items tied to an SccacheServer instance
     pool: tokio::runtime::Handle,
 
@@ -159,7 +158,7 @@ struct DistClientConfig {
 }
 
 #[cfg(feature = "dist-client")]
-enum DistClientState {
+pub enum DistClientState {
     #[cfg(feature = "dist-client")]
     Some(Box<DistClientConfig>, Arc<dyn dist::Client>),
     #[cfg(feature = "dist-client")]
@@ -180,6 +179,11 @@ impl DistClientContainer {
     }
 
     pub fn new_disabled() -> Self {
+        Self {}
+    }
+
+    #[cfg(feature = "dist-client")]
+    pub fn new_with_state(_: DistClientState) -> Self {
         Self {}
     }
 
@@ -208,6 +212,13 @@ impl DistClientContainer {
         };
         let state = Self::create_state(config);
         let state = pool.block_on(state);
+        Self {
+            state: futures::lock::Mutex::new(state),
+        }
+    }
+
+    #[cfg(feature = "dist-client")]
+    pub fn new_with_state(state: DistClientState) -> Self {
         Self {
             state: futures::lock::Mutex::new(state),
         }
@@ -398,15 +409,15 @@ thread_local! {
     /// catch_unwind doesn't provide panic location, so we store that
     /// information via a panic hook to be used when catch_unwind
     /// catches a panic.
-    static PANIC_LOCATION: Cell<Option<(String, u32, u32)>> = Cell::new(None);
+    static PANIC_LOCATION: Cell<Option<(String, u32, u32)>> = const { Cell::new(None) };
 }
 
-/// Start an sccache server, listening on `port`.
+/// Start an sccache server, listening on `addr`.
 ///
 /// Spins an event loop handling client connections until a client
 /// requests a shutdown.
-pub fn start_server(config: &Config, port: u16) -> Result<()> {
-    info!("start_server: port: {}", port);
+pub fn start_server(config: &Config, addr: &crate::net::SocketAddr) -> Result<()> {
+    info!("start_server: {addr}");
     let panic_hook = std::panic::take_hook();
     std::panic::set_hook(Box::new(move |info| {
         PANIC_LOCATION.with(|l| {
@@ -417,7 +428,7 @@ pub fn start_server(config: &Config, port: u16) -> Result<()> {
         });
         panic_hook(info)
     }));
-    let client = unsafe { Client::new() };
+    let client = Client::new();
     let runtime = tokio::runtime::Builder::new_multi_thread()
         .enable_all()
         .worker_threads(std::cmp::max(20, 2 * num_cpus::get()))
@@ -467,59 +478,105 @@ pub fn start_server(config: &Config, port: u16) -> Result<()> {
         _ => raw_storage,
     };
 
-    let res =
-        SccacheServer::<ProcessCommandCreator>::new(port, runtime, client, dist_client, storage);
+    let res = (|| -> io::Result<_> {
+        match addr {
+            crate::net::SocketAddr::Net(addr) => {
+                trace!("binding TCP {addr}");
+                let l = runtime.block_on(tokio::net::TcpListener::bind(addr))?;
+                let srv =
+                    SccacheServer::<_>::with_listener(l, runtime, client, dist_client, storage);
+                Ok((
+                    srv.local_addr().unwrap(),
+                    Box::new(move |f| srv.run(f)) as Box<dyn FnOnce(_) -> _>,
+                ))
+            }
+            #[cfg(unix)]
+            crate::net::SocketAddr::Unix(path) => {
+                trace!("binding unix socket {}", path.display());
+                // Unix socket will report addr in use on any unlink file.
+                let _ = std::fs::remove_file(path);
+                let l = {
+                    let _guard = runtime.enter();
+                    tokio::net::UnixListener::bind(path)?
+                };
+                let srv =
+                    SccacheServer::<_>::with_listener(l, runtime, client, dist_client, storage);
+                Ok((
+                    srv.local_addr().unwrap(),
+                    Box::new(move |f| srv.run(f)) as Box<dyn FnOnce(_) -> _>,
+                ))
+            }
+            #[cfg(any(target_os = "linux", target_os = "android"))]
+            crate::net::SocketAddr::UnixAbstract(p) => {
+                trace!("binding abstract unix socket {}", p.escape_ascii());
+                let abstract_addr = std::os::unix::net::SocketAddr::from_abstract_name(p)?;
+                let l = std::os::unix::net::UnixListener::bind_addr(&abstract_addr)?;
+                l.set_nonblocking(true)?;
+                let l = {
+                    let _guard = runtime.enter();
+                    tokio::net::UnixListener::from_std(l)?
+                };
+                let srv =
+                    SccacheServer::<_>::with_listener(l, runtime, client, dist_client, storage);
+                Ok((
+                    srv.local_addr()
+                        .unwrap_or_else(|| crate::net::SocketAddr::UnixAbstract(p.to_vec())),
+                    Box::new(move |f| srv.run(f)) as Box<dyn FnOnce(_) -> _>,
+                ))
+            }
+        }
+    })();
     match res {
-        Ok(srv) => {
-            let port = srv.port();
-            info!("server started, listening on port {}", port);
-            notify_server_startup(&notify, ServerStartup::Ok { port })?;
-            srv.run(future::pending::<()>())?;
+        Ok((addr, run)) => {
+            info!("server started, listening on {addr}");
+            notify_server_startup(
+                &notify,
+                ServerStartup::Ok {
+                    addr: addr.to_string(),
+                },
+            )?;
+            run(future::pending::<()>())?;
             Ok(())
         }
         Err(e) => {
             error!("failed to start server: {}", e);
-            match e.downcast_ref::<io::Error>() {
-                Some(io_err) if io::ErrorKind::AddrInUse == io_err.kind() => {
-                    notify_server_startup(&notify, ServerStartup::AddrInUse)?;
-                }
-                Some(io_err) if cfg!(windows) && Some(10013) == io_err.raw_os_error() => {
-                    // 10013 is the "WSAEACCES" error, which can occur if the requested port
-                    // has been allocated for other purposes, such as winNAT or Hyper-V.
-                    let windows_help_message =
-                        "A Windows port exclusion is blocking use of the configured port.\nTry setting SCCACHE_SERVER_PORT to a new value.";
-                    let reason: String = format!("{windows_help_message}\n{e}");
-                    notify_server_startup(&notify, ServerStartup::Err { reason })?;
-                }
-                _ => {
-                    let reason = e.to_string();
-                    notify_server_startup(&notify, ServerStartup::Err { reason })?;
-                }
-            };
-            Err(e)
+            if io::ErrorKind::AddrInUse == e.kind() {
+                notify_server_startup(&notify, ServerStartup::AddrInUse)?;
+            } else if cfg!(windows) && Some(10013) == e.raw_os_error() {
+                // 10013 is the "WSAEACCES" error, which can occur if the requested port
+                // has been allocated for other purposes, such as winNAT or Hyper-V.
+                let windows_help_message =
+                    "A Windows port exclusion is blocking use of the configured port.\nTry setting SCCACHE_SERVER_PORT to a new value.";
+                let reason: String = format!("{windows_help_message}\n{e}");
+                notify_server_startup(&notify, ServerStartup::Err { reason })?;
+            } else {
+                let reason = e.to_string();
+                notify_server_startup(&notify, ServerStartup::Err { reason })?;
+            }
+            Err(e.into())
         }
     }
 }
 
-pub struct SccacheServer<C: CommandCreatorSync> {
+pub struct SccacheServer<A: crate::net::Acceptor, C: CommandCreatorSync = ProcessCommandCreator> {
     runtime: Runtime,
-    listener: TcpListener,
+    listener: A,
     rx: mpsc::Receiver<ServerMessage>,
     timeout: Duration,
     service: SccacheService<C>,
     wait: WaitUntilZero,
 }
 
-impl<C: CommandCreatorSync> SccacheServer<C> {
+impl<C: CommandCreatorSync> SccacheServer<tokio::net::TcpListener, C> {
     pub fn new(
         port: u16,
         runtime: Runtime,
         client: Client,
         dist_client: DistClientContainer,
         storage: Arc<dyn Storage>,
-    ) -> Result<SccacheServer<C>> {
-        let addr = SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), port);
-        let listener = runtime.block_on(TcpListener::bind(&SocketAddr::V4(addr)))?;
+    ) -> Result<Self> {
+        let addr = crate::net::SocketAddr::with_port(port);
+        let listener = runtime.block_on(tokio::net::TcpListener::bind(addr.as_net().unwrap()))?;
 
         Ok(Self::with_listener(
             listener,
@@ -529,14 +586,16 @@ impl<C: CommandCreatorSync> SccacheServer<C> {
             storage,
         ))
     }
+}
 
+impl<A: crate::net::Acceptor, C: CommandCreatorSync> SccacheServer<A, C> {
     pub fn with_listener(
-        listener: TcpListener,
+        listener: A,
         runtime: Runtime,
         client: Client,
         dist_client: DistClientContainer,
         storage: Arc<dyn Storage>,
-    ) -> SccacheServer<C> {
+    ) -> Self {
         // Prepare the service which we'll use to service all incoming TCP
         // connections.
         let (tx, rx) = mpsc::channel(1);
@@ -580,8 +639,8 @@ impl<C: CommandCreatorSync> SccacheServer<C> {
 
     /// Returns the port that this server is bound to
     #[allow(dead_code)]
-    pub fn port(&self) -> u16 {
-        self.listener.local_addr().unwrap().port()
+    pub fn local_addr(&self) -> Option<crate::net::SocketAddr> {
+        self.listener.local_addr().unwrap()
     }
 
     /// Runs this server to completion.
@@ -593,6 +652,7 @@ impl<C: CommandCreatorSync> SccacheServer<C> {
     where
         F: Future,
         C: Send,
+        A::Socket: 'static,
     {
         let SccacheServer {
             runtime,
@@ -607,7 +667,7 @@ impl<C: CommandCreatorSync> SccacheServer<C> {
         // connections in separate tasks.
         let server = async move {
             loop {
-                let (socket, _) = listener.accept().await?;
+                let socket = listener.accept().await?;
                 trace!("incoming connection");
                 let conn = service.clone().bind(socket).map_err(|res| {
                     error!("Failed to bind socket: {}", res);
@@ -709,7 +769,7 @@ impl<C> CompilerCacheEntry<C> {
 }
 /// Service implementation for sccache
 #[derive(Clone)]
-struct SccacheService<C>
+pub struct SccacheService<C>
 where
     C: Send,
 {
@@ -756,7 +816,7 @@ where
 }
 
 type SccacheRequest = Message<Request, Body<()>>;
-type SccacheResponse = Message<Response, Body<Response>>;
+type SccacheResponse = Message<Response, Pin<Box<dyn Future<Output = Result<Response>> + Send>>>;
 
 /// Messages sent from all services to the main event loop indicating activity.
 ///
@@ -865,6 +925,60 @@ where
         }
     }
 
+    pub fn mock_with_storage(
+        storage: Arc<dyn Storage>,
+        rt: tokio::runtime::Handle,
+    ) -> SccacheService<C> {
+        let (tx, _) = mpsc::channel(1);
+        let (_, info) = WaitUntilZero::new();
+        let client = Client::new_num(1);
+        let dist_client = DistClientContainer::new_disabled();
+        SccacheService {
+            stats: Arc::default(),
+            dist_client: Arc::new(dist_client),
+            storage,
+            compilers: Arc::default(),
+            compiler_proxies: Arc::default(),
+            rt,
+            creator: C::new(&client),
+            tx,
+            info,
+        }
+    }
+
+    #[cfg(feature = "dist-client")]
+    pub fn mock_with_dist_client(
+        dist_client: Arc<dyn dist::Client>,
+        storage: Arc<dyn Storage>,
+        rt: tokio::runtime::Handle,
+    ) -> SccacheService<C> {
+        let (tx, _) = mpsc::channel(1);
+        let (_, info) = WaitUntilZero::new();
+        let client = Client::new_num(1);
+        SccacheService {
+            stats: Arc::default(),
+            dist_client: Arc::new(DistClientContainer::new_with_state(DistClientState::Some(
+                Box::new(DistClientConfig {
+                    pool: rt.clone(),
+                    scheduler_url: None,
+                    auth: config::DistAuth::Token { token: "".into() },
+                    cache_dir: "".into(),
+                    toolchain_cache_size: 0,
+                    toolchains: vec![],
+                    rewrite_includes_only: false,
+                }),
+                dist_client,
+            ))),
+            storage,
+            compilers: Arc::default(),
+            compiler_proxies: Arc::default(),
+            rt: rt.clone(),
+            creator: C::new(&client),
+            tx,
+            info,
+        }
+    }
+
     fn bind<T>(self, socket: T) -> impl Future<Output = Result<()>> + Send + Sized + 'static
     where
         T: AsyncRead + AsyncWrite + Unpin + Send + 'static,
@@ -889,15 +1003,18 @@ where
         stream
             .err_into::<Error>()
             .and_then(move |input| me.clone().call(input))
-            .and_then(move |message| async move {
-                let fut = match message {
+            .and_then(move |response| async move {
+                let fut = match response {
                     Message::WithoutBody(message) => {
                         let stream = stream::once(async move { Ok(Frame::Message { message }) });
                         Either::Left(stream)
                     }
                     Message::WithBody(message, body) => {
                         let stream = stream::once(async move { Ok(Frame::Message { message }) })
-                            .chain(body.map_ok(|chunk| Frame::Body { chunk: Some(chunk) }))
+                            .chain(
+                                body.into_stream()
+                                    .map_ok(|chunk| Frame::Body { chunk: Some(chunk) }),
+                            )
                             .chain(stream::once(async move { Ok(Frame::Body { chunk: None }) }));
                         Either::Right(stream)
                     }
@@ -944,7 +1061,7 @@ where
 
     /// Look up compiler info from the cache for the compiler `path`.
     /// If not cached, determine the compiler type and cache the result.
-    async fn compiler_info(
+    pub async fn compiler_info(
         &self,
         path: PathBuf,
         cwd: PathBuf,
@@ -1125,11 +1242,17 @@ where
                 match c.parse_arguments(&cmd, &cwd, &env_vars) {
                     CompilerArguments::Ok(hasher) => {
                         debug!("parse_arguments: Ok: {:?}", cmd);
-                        self.stats.lock().await.requests_executed += 1;
-                        let (tx, rx) = Body::pair();
-                        self.start_compile_task(c, hasher, cmd, cwd, env_vars, tx);
-                        let res = CompileResponse::CompileStarted;
-                        return Message::WithBody(Response::Compile(res), rx);
+
+                        let body = self
+                            .clone()
+                            .start_compile_task(c, hasher, cmd, cwd, env_vars)
+                            .and_then(|res| async { Ok(Response::CompileFinished(res)) })
+                            .boxed();
+
+                        return Message::WithBody(
+                            Response::Compile(CompileResponse::CompileStarted),
+                            body,
+                        );
                     }
                     CompilerArguments::CannotCache(why, extra_info) => {
                         if let Some(extra_info) = extra_info {
@@ -1159,192 +1282,234 @@ where
     /// Given compiler arguments `arguments`, look up
     /// a compile result in the cache or execute the compilation and store
     /// the result in the cache.
-    fn start_compile_task(
-        &self,
+    pub async fn start_compile_task(
+        self,
         compiler: Box<dyn Compiler<C>>,
         hasher: Box<dyn CompilerHasher<C>>,
         arguments: Vec<OsString>,
         cwd: PathBuf,
         env_vars: Vec<(OsString, OsString)>,
-        mut tx: mpsc::Sender<Result<Response>>,
-    ) {
-        let force_recache = env_vars
-            .iter()
-            .any(|(k, _v)| k.as_os_str() == OsStr::new("SCCACHE_RECACHE"));
-        let cache_control = if force_recache {
+    ) -> Result<CompileFinished> {
+        self.stats.lock().await.requests_executed += 1;
+
+        let force_recache = env_vars.iter().any(|(k, _v)| k == "SCCACHE_RECACHE");
+        let force_no_cache = env_vars.iter().any(|(k, _v)| k == "SCCACHE_NO_CACHE");
+
+        let cache_control = if force_no_cache {
+            CacheControl::ForceNoCache
+        } else if force_recache {
             CacheControl::ForceRecache
         } else {
             CacheControl::Default
         };
+
         let out_pretty = hasher.output_pretty().into_owned();
         let color_mode = hasher.color_mode();
+
+        let (kind, lang) = {
+            // HACK: See note in src/compiler/nvcc.rs
+            if env_vars
+                .iter()
+                .any(|(k, _)| k == "__SCCACHE_THIS_IS_A_CUDA_COMPILATION__")
+            {
+                (
+                    CompilerKind::C(crate::compiler::CCompilerKind::Nvcc),
+                    Language::Cuda,
+                )
+            } else {
+                (compiler.kind(), hasher.language())
+            }
+        };
+
         let me = self.clone();
-        let kind = compiler.kind();
-        let lang = hasher.language();
-        let creator = self.creator.clone();
-        let storage = self.storage.clone();
-        let pool = self.rt.clone();
 
-        let task = async move {
-            let dist_client = me.dist_client.get_client().await;
-            let result = match dist_client {
-                Ok(client) => std::panic::AssertUnwindSafe(hasher.get_cached_or_compile(
-                    client,
-                    creator,
-                    storage,
-                    arguments,
-                    cwd,
-                    env_vars,
-                    cache_control,
-                    pool,
-                ))
-                .catch_unwind()
-                .await
-                .map_err(|e| {
-                    let panic = e
-                        .downcast_ref::<&str>()
-                        .map(|s| &**s)
-                        .or_else(|| e.downcast_ref::<String>().map(|s| &**s))
-                        .unwrap_or("An unknown panic was caught.");
-                    let thread = std::thread::current();
-                    let thread_name = thread.name().unwrap_or("unnamed");
-                    if let Some((file, line, column)) = PANIC_LOCATION.with(|l| l.take()) {
-                        anyhow!(
-                            "thread '{thread_name}' panicked at {file}:{line}:{column}: {panic}"
+        self.rt
+            .spawn(async move {
+
+                let result = match me.dist_client.get_client().await {
+                    Ok(client) => {
+                        std::panic::AssertUnwindSafe(hasher
+                            .get_cached_or_compile(
+                                &me,
+                                client,
+                                me.creator.clone(),
+                                me.storage.clone(),
+                                arguments,
+                                cwd,
+                                env_vars,
+                                cache_control,
+                                me.rt.clone(),
+                            )
                         )
-                    } else {
-                        anyhow!("thread '{thread_name}' panicked: {panic}")
+                        .catch_unwind()
+                        .await
+                        .map_err(|e| {
+                            let panic = e
+                                .downcast_ref::<&str>()
+                                .map(|s| &**s)
+                                .or_else(|| e.downcast_ref::<String>().map(|s| &**s))
+                                .unwrap_or("An unknown panic was caught.");
+                            let thread = std::thread::current();
+                            let thread_name = thread.name().unwrap_or("unnamed");
+                            if let Some((file, line, column)) = PANIC_LOCATION.with(|l| l.take()) {
+                                anyhow!("thread '{thread_name}' panicked at {file}:{line}:{column}: {panic}")
+                            } else {
+                                anyhow!("thread '{thread_name}' panicked: {panic}")
+                            }
+                        })
+                        .and_then(std::convert::identity)
                     }
-                })
-                .and_then(std::convert::identity),
-                Err(e) => Err(e),
-            };
-            let mut cache_write = None;
-            let mut res = CompileFinished {
-                color_mode,
-                ..Default::default()
-            };
-            match result {
-                Ok((compiled, out)) => {
-                    let mut stats = me.stats.lock().await;
-                    match compiled {
-                        CompileResult::Error => {
-                            debug!("compile result: cache error");
+                    Err(e) => Err(e),
+                };
 
-                            stats.cache_errors.increment(&kind, &lang);
-                        }
-                        CompileResult::CacheHit(duration) => {
-                            debug!("compile result: cache hit");
+                let mut cache_write = None;
+                let mut res = CompileFinished {
+                    color_mode,
+                    ..Default::default()
+                };
 
-                            stats.cache_hits.increment(&kind, &lang);
-                            stats.cache_read_hit_duration += duration;
-                        }
-                        CompileResult::CacheMiss(miss_type, dist_type, duration, future) => {
-                            debug!("compile result: cache miss");
+                let mut stats = me.stats.lock().await;
 
-                            match dist_type {
-                                DistType::NoDist => {}
-                                DistType::Ok(id) => {
-                                    let server = id.addr().to_string();
-                                    let server_count =
-                                        stats.dist_compiles.entry(server).or_insert(0);
-                                    *server_count += 1;
-                                }
-                                DistType::Error => stats.dist_errors += 1,
-                            }
-                            match miss_type {
-                                MissType::Normal => {}
-                                MissType::ForcedRecache => {
-                                    stats.forced_recaches += 1;
-                                }
-                                MissType::TimedOut => {
-                                    stats.cache_timeouts += 1;
-                                }
-                                MissType::CacheReadError => {
-                                    stats.cache_errors.increment(&kind, &lang);
-                                }
-                            }
-                            stats.cache_misses.increment(&kind, &lang);
-                            stats.compiler_write_duration += duration;
-                            debug!("stats after compile result: {stats:?}");
-                            cache_write = Some(future);
-                        }
-                        CompileResult::NotCacheable => {
-                            debug!("compile result: not cacheable");
+                match result {
+                    Ok((compiled, out)) => {
 
-                            stats.cache_misses.increment(&kind, &lang);
-                            stats.non_cacheable_compilations += 1;
-                        }
-                        CompileResult::CompileFailed => {
-                            debug!("compile result: compile failed");
+                        let mut dist_type = DistType::NoDist;
 
-                            stats.compile_fails += 1;
-                        }
-                    };
-                    // Make sure the write guard has been dropped ASAP.
-                    drop(stats);
+                        match compiled {
+                            CompileResult::Error => {
+                                debug!("compile result: cache error");
 
-                    let Output {
-                        status,
-                        stdout,
-                        stderr,
-                    } = out;
-                    trace!("CompileFinished retcode: {}", status);
-                    match status.code() {
-                        Some(code) => res.retcode = Some(code),
-                        None => res.signal = Some(get_signal(status)),
-                    };
-                    res.stdout = stdout;
-                    res.stderr = stderr;
-                }
-                Err(err) => {
-                    let mut stats = me.stats.lock().await;
-                    match err.downcast::<ProcessError>() {
-                        Ok(ProcessError(output)) => {
-                            debug!("Compilation failed: {:?}", output);
-                            stats.compile_fails += 1;
-                            match output.status.code() {
-                                Some(code) => res.retcode = Some(code),
-                                None => res.signal = Some(get_signal(output.status)),
-                            };
-                            res.stdout = output.stdout;
-                            res.stderr = output.stderr;
-                        }
-                        Err(err) => match err.downcast::<HttpClientError>() {
-                            Ok(HttpClientError(msg)) => {
-                                me.dist_client.reset_state().await;
-                                let errmsg =
-                                    format!("[{:?}] http error status: {}", out_pretty, msg);
-                                error!("{}", errmsg);
-                                res.retcode = Some(1);
-                                res.stderr = errmsg.as_bytes().to_vec();
-                            }
-                            Err(err) => {
-                                use std::fmt::Write;
-
-                                error!("[{:?}] fatal error: {}", out_pretty, err);
-
-                                let mut error = "sccache: encountered fatal error\n".to_string();
-                                let _ = writeln!(error, "sccache: error: {}", err);
-                                for e in err.chain() {
-                                    error!("[{:?}] \t{}", out_pretty, e);
-                                    let _ = writeln!(error, "sccache: caused by: {}", e);
-                                }
                                 stats.cache_errors.increment(&kind, &lang);
-                                //TODO: figure out a better way to communicate this?
-                                res.retcode = Some(-2);
-                                res.stderr = error.into_bytes();
                             }
-                        },
-                    }
-                }
-            };
-            let send = tx
-                .send(Ok(Response::CompileFinished(res)))
-                .map_err(|e| anyhow!("send on finish failed").context(e));
+                            CompileResult::CacheHit(duration) => {
+                                debug!("compile result: cache hit");
 
-            let me = me.clone();
-            let cache_write = async move {
+                                stats.cache_hits.increment(&kind, &lang);
+                                stats.cache_read_hit_duration += duration;
+                            }
+                            CompileResult::CacheMiss(miss_type, dt, duration, future) => {
+                                debug!("[{}]: compile result: cache miss", out_pretty);
+                                dist_type = dt;
+
+                                match miss_type {
+                                    MissType::Normal => {}
+                                    MissType::ForcedNoCache => {}
+                                    MissType::ForcedRecache => {
+                                        stats.forced_recaches += 1;
+                                    }
+                                    MissType::TimedOut => {
+                                        stats.cache_timeouts += 1;
+                                    }
+                                    MissType::CacheReadError => {
+                                        stats.cache_errors.increment(&kind, &lang);
+                                    }
+                                }
+                                stats.compilations += 1;
+                                stats.cache_misses.increment(&kind, &lang);
+                                stats.compiler_write_duration += duration;
+                                debug!("stats after compile result: {stats:?}");
+                                cache_write = Some(future);
+                            }
+                            CompileResult::NotCached(dt, duration) => {
+                                debug!("[{}]: compile result: not cached", out_pretty);
+                                dist_type = dt;
+                                stats.compilations += 1;
+                                stats.compiler_write_duration += duration;
+                            }
+                            CompileResult::NotCacheable(dt, duration) => {
+                                debug!("[{}]: compile result: not cacheable", out_pretty);
+                                dist_type = dt;
+                                stats.compilations += 1;
+                                stats.compiler_write_duration += duration;
+                                stats.non_cacheable_compilations += 1;
+                            }
+                            CompileResult::CompileFailed(dt, duration) => {
+                                debug!("[{}]: compile result: compile failed", out_pretty);
+                                dist_type = dt;
+                                stats.compilations += 1;
+                                stats.compiler_write_duration += duration;
+                                stats.compile_fails += 1;
+                            }
+                        };
+
+                        match dist_type {
+                            DistType::NoDist => {}
+                            DistType::Ok(id) => {
+                                let server = id.addr().to_string();
+                                let server_count = stats.dist_compiles.entry(server).or_insert(0);
+                                *server_count += 1;
+                            }
+                            DistType::Error => stats.dist_errors += 1,
+                        }
+
+                        // Make sure the write guard has been dropped ASAP.
+                        drop(stats);
+
+                        let Output {
+                            status,
+                            stdout,
+                            stderr,
+                        } = out;
+
+                        trace!("CompileFinished retcode: {}", status);
+
+                        match status.code() {
+                            Some(code) => res.retcode = Some(code),
+                            None => res.signal = Some(get_signal(status)),
+                        };
+
+                        res.stdout = stdout;
+                        res.stderr = stderr;
+                    }
+                    Err(err) => {
+                        match err.downcast::<ProcessError>() {
+                            Ok(ProcessError(output)) => {
+                                debug!("Compilation failed: {:?}", output);
+                                stats.compile_fails += 1;
+                                // Make sure the write guard has been dropped ASAP.
+                                drop(stats);
+
+                                match output.status.code() {
+                                    Some(code) => res.retcode = Some(code),
+                                    None => res.signal = Some(get_signal(output.status)),
+                                };
+                                res.stdout = output.stdout;
+                                res.stderr = output.stderr;
+                            }
+                            Err(err) => match err.downcast::<HttpClientError>() {
+                                Ok(HttpClientError(msg)) => {
+                                    // Make sure the write guard has been dropped ASAP.
+                                    drop(stats);
+                                    me.dist_client.reset_state().await;
+                                    let errmsg = format!("[{:?}] http error status: {}", out_pretty, msg);
+                                    error!("{}", errmsg);
+                                    res.retcode = Some(1);
+                                    res.stderr = errmsg.as_bytes().to_vec();
+                                }
+                                Err(err) => {
+                                    stats.cache_errors.increment(&kind, &lang);
+                                    // Make sure the write guard has been dropped ASAP.
+                                    drop(stats);
+
+                                    use std::fmt::Write;
+
+                                    error!("[{:?}] fatal error: {}", out_pretty, err);
+
+                                    let mut error = "sccache: encountered fatal error\n".to_string();
+                                    let _ = writeln!(error, "sccache: error: {}", err);
+                                    for e in err.chain() {
+                                        error!("[{:?}] \t{}", out_pretty, e);
+                                        let _ = writeln!(error, "sccache: caused by: {}", e);
+                                    }
+                                    //TODO: figure out a better way to communicate this?
+                                    res.retcode = Some(-2);
+                                    res.stderr = error.into_bytes();
+                                }
+                            },
+                        }
+                    }
+                };
+
                 if let Some(cache_write) = cache_write {
                     match cache_write.await {
                         Err(e) => {
@@ -1364,18 +1529,11 @@ where
                         }
                     }
                 }
-                Ok(())
-            };
 
-            futures::future::try_join(send, cache_write).await?;
-
-            Ok::<_, Error>(())
-        };
-
-        self.rt.spawn(async move {
-            task.await
-                .unwrap_or_else(|e| warn!("Failed to execute task: {:?}", e));
-        });
+                Ok(res)
+            })
+            .map_err(anyhow::Error::new)
+            .await?
     }
 }
 
@@ -1448,7 +1606,9 @@ pub struct ServerStats {
     pub cache_write_duration: Duration,
     /// The total time spent reading cache hits.
     pub cache_read_hit_duration: Duration,
-    /// The total time spent reading cache misses.
+    /// The number of compilations performed.
+    pub compilations: u64,
+    /// The total time spent compiling.
     pub compiler_write_duration: Duration,
     /// The count of compilation failures.
     pub compile_fails: u64,
@@ -1501,6 +1661,7 @@ impl Default for ServerStats {
             cache_writes: u64::default(),
             cache_write_duration: Duration::new(0, 0),
             cache_read_hit_duration: Duration::new(0, 0),
+            compilations: u64::default(),
             compiler_write_duration: Duration::new(0, 0),
             compile_fails: u64::default(),
             not_cached: HashMap::new(),
@@ -1510,11 +1671,23 @@ impl Default for ServerStats {
     }
 }
 
+pub trait ServerStatsWriter {
+    fn write(&mut self, text: &str);
+}
+
+pub struct StdoutServerStatsWriter;
+
+impl ServerStatsWriter for StdoutServerStatsWriter {
+    fn write(&mut self, text: &str) {
+        println!("{text}");
+    }
+}
+
 impl ServerStats {
-    /// Print stats to stdout in a human-readable format.
+    /// Print stats in a human-readable format.
     ///
     /// Return the formatted width of each of the (name, value) columns.
-    fn print(&self, advanced: bool) -> (usize, usize) {
+    fn print<T: ServerStatsWriter>(&self, writer: &mut T, advanced: bool) -> (usize, usize) {
         macro_rules! set_stat {
             ($vec:ident, $var:expr, $name:expr) => {{
                 // name, value, suffix length
@@ -1570,16 +1743,21 @@ impl ServerStats {
             set_lang_stat!(stats_vec, self.cache_hits, "Cache hits");
             set_lang_stat!(stats_vec, self.cache_misses, "Cache misses");
         }
+
+        self.set_percentage_stats(&mut stats_vec, advanced);
+
         set_stat!(stats_vec, self.cache_timeouts, "Cache timeouts");
         set_stat!(stats_vec, self.cache_read_errors, "Cache read errors");
         set_stat!(stats_vec, self.forced_recaches, "Forced recaches");
         set_stat!(stats_vec, self.cache_write_errors, "Cache write errors");
-        set_stat!(stats_vec, self.compile_fails, "Compilation failures");
         if advanced {
             set_compiler_stat!(stats_vec, self.cache_errors, "Cache errors");
         } else {
             set_lang_stat!(stats_vec, self.cache_errors, "Cache errors");
         }
+
+        set_stat!(stats_vec, self.compilations, "Compilations");
+        set_stat!(stats_vec, self.compile_fails, "Compilation failures");
 
         set_stat!(
             stats_vec,
@@ -1610,7 +1788,7 @@ impl ServerStats {
         set_duration_stat!(
             stats_vec,
             self.compiler_write_duration,
-            self.cache_misses.all(),
+            self.compilations,
             "Average compiler"
         );
         set_duration_stat!(
@@ -1627,44 +1805,109 @@ impl ServerStats {
         let name_width = stats_vec.iter().map(|(n, _, _)| n.len()).max().unwrap();
         let stat_width = stats_vec.iter().map(|(_, s, _)| s.len()).max().unwrap();
         for (name, stat, suffix_len) in stats_vec {
-            println!(
+            writer.write(&format!(
                 "{:<name_width$} {:>stat_width$}",
                 name,
                 stat,
                 name_width = name_width,
                 stat_width = stat_width + suffix_len
-            );
+            ));
         }
         if !self.dist_compiles.is_empty() {
-            println!("\nSuccessful distributed compiles");
+            writer.write("\nSuccessful distributed compiles");
             let mut counts: Vec<_> = self.dist_compiles.iter().collect();
             counts.sort_by(|(_, c1), (_, c2)| c1.cmp(c2).reverse());
             for (reason, count) in counts {
-                println!(
+                writer.write(&format!(
                     "  {:<name_width$} {:>stat_width$}",
                     reason,
                     count,
                     name_width = name_width - 2,
-                    stat_width = stat_width
-                );
+                    stat_width = stat_width,
+                ));
             }
         }
         if !self.not_cached.is_empty() {
-            println!("\nNon-cacheable reasons:");
+            writer.write("\nNon-cacheable reasons:");
             let mut counts: Vec<_> = self.not_cached.iter().collect();
             counts.sort_by(|(_, c1), (_, c2)| c1.cmp(c2).reverse());
             for (reason, count) in counts {
-                println!(
+                writer.write(&format!(
                     "{:<name_width$} {:>stat_width$}",
                     reason,
                     count,
                     name_width = name_width,
-                    stat_width = stat_width
-                );
+                    stat_width = stat_width,
+                ));
             }
-            println!();
+            writer.write("");
         }
         (name_width, stat_width)
+    }
+
+    fn set_percentage_stats(&self, stats_vec: &mut Vec<(String, String, usize)>, advanced: bool) {
+        set_percentage_stat(
+            stats_vec,
+            self.cache_hits.all(),
+            self.cache_misses.all() + self.cache_hits.all(),
+            "Cache hits rate",
+        );
+
+        let (stats_hits, stats_misses): (Vec<_>, Vec<_>) = if advanced {
+            (
+                self.cache_hits.adv_counts.iter().collect(),
+                self.cache_misses.adv_counts.iter().collect(),
+            )
+        } else {
+            (
+                self.cache_hits.counts.iter().collect(),
+                self.cache_misses.counts.iter().collect(),
+            )
+        };
+
+        let mut all_languages: HashSet<&String> = HashSet::new();
+        for (lang, _) in &stats_hits {
+            all_languages.insert(lang);
+        }
+        for (lang, _) in &stats_misses {
+            all_languages.insert(lang);
+        }
+
+        let mut all_languages: Vec<&String> = all_languages.into_iter().collect();
+        all_languages.sort();
+
+        for lang in all_languages {
+            let count_hits = stats_hits
+                .iter()
+                .find(|&&(l, _)| l == lang)
+                .map_or(0, |&(_, &count)| count);
+
+            let count_misses = stats_misses
+                .iter()
+                .find(|&&(l, _)| l == lang)
+                .map_or(0, |&(_, &count)| count);
+
+            set_percentage_stat(
+                stats_vec,
+                count_hits,
+                count_hits + count_misses,
+                &format!("Cache hits rate ({})", lang),
+            );
+        }
+    }
+}
+
+fn set_percentage_stat(
+    vec: &mut Vec<(String, String, usize)>,
+    count_hits: u64,
+    total: u64,
+    name: &str,
+) {
+    if total == 0 {
+        vec.push((name.to_string(), "-".to_string(), 0));
+    } else {
+        let ratio = count_hits as f64 / total as f64;
+        vec.push((name.to_string(), format!("{:.2} %", ratio * 100.0), 2));
     }
 }
 
@@ -1700,7 +1943,7 @@ impl ServerInfo {
 
     /// Print info to stdout in a human-readable format.
     pub fn print(&self, advanced: bool) {
-        let (name_width, stat_width) = self.stats.print(advanced);
+        let (name_width, stat_width) = self.stats.print(&mut StdoutServerStatsWriter, advanced);
         println!(
             "{:<name_width$} {}",
             "Cache location",
@@ -1756,13 +1999,6 @@ enum Frame<R, R1> {
 
 struct Body<R> {
     receiver: mpsc::Receiver<Result<R>>,
-}
-
-impl<R> Body<R> {
-    fn pair() -> (mpsc::Sender<Result<R>>, Self) {
-        let (tx, rx) = mpsc::channel(0);
-        (tx, Body { receiver: rx })
-    }
 }
 
 impl<R> futures::Stream for Body<R> {
@@ -1919,7 +2155,7 @@ struct WaitUntilZero {
 
 #[derive(Clone)]
 #[allow(dead_code)]
-struct ActiveInfo {
+pub struct ActiveInfo {
     info: Arc<std::sync::Mutex<Info>>,
 }
 
@@ -1977,4 +2213,109 @@ fn waits_until_zero() {
     drop(active);
     drop(active2);
     assert_eq!(wait.now_or_never(), Some(()));
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    struct StringWriter {
+        buffer: String,
+    }
+
+    impl StringWriter {
+        fn new() -> StringWriter {
+            StringWriter {
+                buffer: String::new(),
+            }
+        }
+
+        fn get_output(self) -> String {
+            self.buffer
+        }
+    }
+
+    impl ServerStatsWriter for StringWriter {
+        fn write(&mut self, text: &str) {
+            self.buffer.push_str(&format!("{}\n", text));
+        }
+    }
+
+    #[test]
+    fn test_print_cache_hits_rate_default_server_stats() {
+        let stats = ServerStats::default();
+
+        let mut writer = StringWriter::new();
+        stats.print(&mut writer, false);
+
+        let output = writer.get_output();
+
+        assert!(output.contains("Cache hits rate                       -"));
+    }
+
+    #[test]
+    fn test_print_cache_hits_rate_server_stats() {
+        let mut cache_hits_counts = HashMap::new();
+        cache_hits_counts.insert("Rust".to_string(), 100);
+        cache_hits_counts.insert("C/C++".to_string(), 200);
+
+        let mut cache_misses_counts = HashMap::new();
+        cache_misses_counts.insert("Rust".to_string(), 50);
+        cache_misses_counts.insert("Cuda".to_string(), 300);
+
+        let stats = ServerStats {
+            cache_hits: PerLanguageCount {
+                counts: cache_hits_counts,
+                ..Default::default()
+            },
+            cache_misses: PerLanguageCount {
+                counts: cache_misses_counts,
+                ..Default::default()
+            },
+            ..Default::default()
+        };
+
+        let mut writer = StringWriter::new();
+        stats.print(&mut writer, false);
+
+        let output = writer.get_output();
+
+        assert!(output.contains("Cache hits rate                    46.15 %"));
+        assert!(output.contains("Cache hits rate (C/C++)           100.00 %"));
+        assert!(output.contains("Cache hits rate (Cuda)              0.00 %"));
+        assert!(output.contains("Cache hits rate (Rust)             66.67 %"));
+    }
+
+    #[test]
+    fn test_print_cache_hits_rate_advanced_server_stats() {
+        let mut cache_hits_counts = HashMap::new();
+        cache_hits_counts.insert("rust".to_string(), 50);
+        cache_hits_counts.insert("c/c++ [clang]".to_string(), 30);
+
+        let mut cache_misses_counts = HashMap::new();
+        cache_misses_counts.insert("rust".to_string(), 100);
+        cache_misses_counts.insert("cuda".to_string(), 70);
+
+        let stats = ServerStats {
+            cache_hits: PerLanguageCount {
+                adv_counts: cache_hits_counts,
+                ..Default::default()
+            },
+            cache_misses: PerLanguageCount {
+                adv_counts: cache_misses_counts,
+                ..Default::default()
+            },
+            ..Default::default()
+        };
+
+        let mut writer = StringWriter::new();
+        stats.print(&mut writer, true);
+
+        let output = writer.get_output();
+
+        assert!(output.contains("Cache hits rate                        -"));
+        assert!(output.contains("Cache hits rate (c/c++ [clang])   100.00 %"));
+        assert!(output.contains("Cache hits rate (cuda)              0.00 %"));
+        assert!(output.contains("Cache hits rate (rust)             33.33 %"));
+    }
 }
