@@ -19,24 +19,27 @@ use opendal::Operator;
 use crate::errors::*;
 use crate::VERSION;
 
+use super::http_client::set_user_agent;
+
 /// A cache that stores entries in GHA Cache Services.
 pub struct GHACache;
 
 impl GHACache {
     pub fn build(version: &str) -> Result<Operator> {
-        let mut builder = Ghac::default();
-        // This is the prefix of gha cache.
-        // From user side, cache key will be like `sccache/f/c/b/fcbxxx`
-        //
-        // User customization is theoretically supported, but I decided
-        // to see the community feedback first.
-        builder.root("/sccache");
+        let mut builder = Ghac::default()
+            // This is the prefix of gha cache.
+            // From user side, cache key will be like `sccache/f/c/b/fcbxxx`
+            //
+            // User customization is theoretically supported, but I decided
+            // to see the community feedback first.
+            .root("/sccache")
+            .http_client(set_user_agent());
 
-        if version.is_empty() {
-            builder.version(&format!("sccache-v{VERSION}"));
+        builder = if version.is_empty() {
+            builder.version(&format!("sccache-v{VERSION}"))
         } else {
-            builder.version(&format!("sccache-v{VERSION}-{version}"));
-        }
+            builder.version(&format!("sccache-v{VERSION}-{version}"))
+        };
 
         let op = Operator::new(builder)?
             .layer(LoggingLayer::default())
