@@ -13,28 +13,28 @@
 // limitations under the License.
 
 use crate::errors::*;
+use crate::net::Connection;
 use crate::protocol::{Request, Response};
 use crate::util;
 use byteorder::{BigEndian, ByteOrder};
 use retry::{delay::Fixed, retry};
 use std::io::{self, BufReader, BufWriter, Read};
-use std::net::TcpStream;
 
 /// A connection to an sccache server.
 pub struct ServerConnection {
     /// A reader for the socket connected to the server.
-    reader: BufReader<TcpStream>,
+    reader: BufReader<Box<dyn Connection>>,
     /// A writer for the socket connected to the server.
-    writer: BufWriter<TcpStream>,
+    writer: BufWriter<Box<dyn Connection>>,
 }
 
 impl ServerConnection {
     /// Create a new connection using `stream`.
-    pub fn new(stream: TcpStream) -> io::Result<ServerConnection> {
-        let writer = stream.try_clone()?;
+    pub fn new(conn: Box<dyn Connection>) -> io::Result<ServerConnection> {
+        let write_conn = conn.try_clone()?;
         Ok(ServerConnection {
-            reader: BufReader::new(stream),
-            writer: BufWriter::new(writer),
+            reader: BufReader::new(conn),
+            writer: BufWriter::new(write_conn),
         })
     }
 
@@ -62,24 +62,24 @@ impl ServerConnection {
     }
 }
 
-/// Establish a TCP connection to an sccache server listening on `port`.
-pub fn connect_to_server(port: u16) -> io::Result<ServerConnection> {
-    trace!("connect_to_server({})", port);
-    let stream = TcpStream::connect(("127.0.0.1", port))?;
-    ServerConnection::new(stream)
+/// Establish a TCP connection to an sccache server listening on `addr`.
+pub fn connect_to_server(addr: &crate::net::SocketAddr) -> io::Result<ServerConnection> {
+    trace!("connect_to_server({addr})");
+    let conn = crate::net::connect(addr)?;
+    ServerConnection::new(conn)
 }
 
-/// Attempt to establish a TCP connection to an sccache server listening on `port`.
+/// Attempt to establish a TCP connection to an sccache server listening on `addr`.
 ///
 /// If the connection fails, retry a few times.
-pub fn connect_with_retry(port: u16) -> io::Result<ServerConnection> {
-    trace!("connect_with_retry({})", port);
+pub fn connect_with_retry(addr: &crate::net::SocketAddr) -> io::Result<ServerConnection> {
+    trace!("connect_with_retry({addr})");
     // TODOs:
     // * Pass the server Child in here, so we can stop retrying
     //   if the process exited.
     // * Send a pipe handle to the server process so it can notify
     //   us once it starts the server instead of us polling.
-    match retry(Fixed::from_millis(500).take(10), || connect_to_server(port)) {
+    match retry(Fixed::from_millis(500).take(10), || connect_to_server(addr)) {
         Ok(conn) => Ok(conn),
         Err(e) => Err(io::Error::new(
             io::ErrorKind::TimedOut,
