@@ -16,6 +16,7 @@
 use crate::cache::{Cache, CacheWrite, DecompressionFailure, FileObjectSource, Storage};
 use crate::compiler::args::*;
 use crate::compiler::c::{CCompiler, CCompilerKind};
+use crate::compiler::cicc::Cicc;
 use crate::compiler::clang::Clang;
 use crate::compiler::diab::Diab;
 use crate::compiler::gcc::Gcc;
@@ -24,6 +25,7 @@ use crate::compiler::msvc::Msvc;
 use crate::compiler::nvcc::Nvcc;
 use crate::compiler::nvcc::NvccHostCompiler;
 use crate::compiler::nvhpc::Nvhpc;
+use crate::compiler::ptxas::Ptxas;
 use crate::compiler::rust::{Rust, RustupProxy};
 use crate::compiler::tasking_vx::TaskingVX;
 #[cfg(feature = "dist-client")]
@@ -31,7 +33,8 @@ use crate::dist::pkg;
 #[cfg(feature = "dist-client")]
 use crate::lru_disk_cache;
 use crate::mock_command::{exit_status, CommandChild, CommandCreatorSync, RunCommand};
-use crate::util::{fmt_duration_as_secs, ref_env, run_input_output};
+use crate::server;
+use crate::util::{fmt_duration_as_secs, run_input_output};
 use crate::{counted_array, dist};
 use async_trait::async_trait;
 use filetime::FileTime;
@@ -71,24 +74,130 @@ pub const CAN_DIST_DYLIBS: bool = true;
 ))]
 pub const CAN_DIST_DYLIBS: bool = false;
 
-#[derive(Clone, Debug)]
-pub struct CompileCommand {
+#[async_trait]
+pub trait CompileCommand<T>: Send + Sync + 'static
+where
+    T: CommandCreatorSync,
+{
+    async fn execute(
+        &self,
+        service: &server::SccacheService<T>,
+        creator: &T,
+    ) -> Result<process::Output>;
+
+    fn get_executable(&self) -> PathBuf;
+    fn get_arguments(&self) -> Vec<OsString>;
+    fn get_env_vars(&self) -> Vec<(OsString, OsString)>;
+    fn get_cwd(&self) -> PathBuf;
+}
+
+#[derive(Debug)]
+pub struct CCompileCommand<I>
+where
+    I: CompileCommandImpl,
+{
+    cmd: I,
+}
+
+impl<I> CCompileCommand<I>
+where
+    I: CompileCommandImpl,
+{
+    #[allow(clippy::new_ret_no_self)]
+    pub fn new<T>(cmd: I) -> Box<dyn CompileCommand<T>>
+    where
+        T: CommandCreatorSync,
+    {
+        Box::new(CCompileCommand { cmd }) as Box<dyn CompileCommand<T>>
+    }
+}
+
+#[async_trait]
+impl<T, I> CompileCommand<T> for CCompileCommand<I>
+where
+    T: CommandCreatorSync,
+    I: CompileCommandImpl,
+{
+    fn get_executable(&self) -> PathBuf {
+        self.cmd.get_executable()
+    }
+    fn get_arguments(&self) -> Vec<OsString> {
+        self.cmd.get_arguments()
+    }
+    fn get_env_vars(&self) -> Vec<(OsString, OsString)> {
+        self.cmd.get_env_vars()
+    }
+    fn get_cwd(&self) -> PathBuf {
+        self.cmd.get_cwd()
+    }
+
+    async fn execute(
+        &self,
+        service: &server::SccacheService<T>,
+        creator: &T,
+    ) -> Result<process::Output> {
+        self.cmd.execute(service, creator).await
+    }
+}
+
+#[async_trait]
+pub trait CompileCommandImpl: Send + Sync + 'static {
+    fn get_executable(&self) -> PathBuf;
+    fn get_arguments(&self) -> Vec<OsString>;
+    fn get_env_vars(&self) -> Vec<(OsString, OsString)>;
+    fn get_cwd(&self) -> PathBuf;
+
+    async fn execute<T>(
+        &self,
+        service: &server::SccacheService<T>,
+        creator: &T,
+    ) -> Result<process::Output>
+    where
+        T: CommandCreatorSync;
+}
+
+#[derive(Debug)]
+pub struct SingleCompileCommand {
     pub executable: PathBuf,
     pub arguments: Vec<OsString>,
     pub env_vars: Vec<(OsString, OsString)>,
     pub cwd: PathBuf,
 }
 
-impl CompileCommand {
-    pub async fn execute<T>(self, creator: &T) -> Result<process::Output>
+#[async_trait]
+impl CompileCommandImpl for SingleCompileCommand {
+    fn get_executable(&self) -> PathBuf {
+        self.executable.clone()
+    }
+    fn get_arguments(&self) -> Vec<OsString> {
+        self.arguments.clone()
+    }
+    fn get_env_vars(&self) -> Vec<(OsString, OsString)> {
+        self.env_vars.clone()
+    }
+    fn get_cwd(&self) -> PathBuf {
+        self.cwd.clone()
+    }
+
+    async fn execute<T>(
+        &self,
+        _: &server::SccacheService<T>,
+        creator: &T,
+    ) -> Result<process::Output>
     where
         T: CommandCreatorSync,
     {
-        let mut cmd = creator.clone().new_command_sync(self.executable);
-        cmd.args(&self.arguments)
+        let SingleCompileCommand {
+            executable,
+            arguments,
+            env_vars,
+            cwd,
+        } = self;
+        let mut cmd = creator.clone().new_command_sync(executable);
+        cmd.args(arguments)
             .env_clear()
-            .envs(self.env_vars)
-            .current_dir(self.cwd);
+            .envs(env_vars.to_vec())
+            .current_dir(cwd);
         run_input_output(cmd, None).await
     }
 }
@@ -111,7 +220,10 @@ pub enum Language {
     CxxHeader,
     ObjectiveC,
     ObjectiveCxx,
+    ObjectiveCxxHeader,
     Cuda,
+    Ptx,
+    Cubin,
     Rust,
     Hip,
 }
@@ -134,6 +246,8 @@ impl Language {
             Some("M") | Some("mm") => Some(Language::ObjectiveCxx),
             // TODO mii
             Some("cu") => Some(Language::Cuda),
+            Some("ptx") => Some(Language::Ptx),
+            Some("cubin") => Some(Language::Cubin),
             // TODO cy
             Some("rs") => Some(Language::Rust),
             Some("hip") => Some(Language::Hip),
@@ -146,12 +260,16 @@ impl Language {
 
     pub fn as_str(self) -> &'static str {
         match self {
-            Language::C | Language::CHeader => "c",
-            Language::Cxx | Language::CxxHeader => "c++",
+            Language::C => "c",
+            Language::CHeader => "cHeader",
+            Language::Cxx => "c++",
+            Language::CxxHeader => "c++Header",
             Language::GenericHeader => "c/c++",
             Language::ObjectiveC => "objc",
-            Language::ObjectiveCxx => "objc++",
+            Language::ObjectiveCxx | Language::ObjectiveCxxHeader => "objc++",
             Language::Cuda => "cuda",
+            Language::Ptx => "ptx",
+            Language::Cubin => "cubin",
             Language::Rust => "rust",
             Language::Hip => "hip",
         }
@@ -167,8 +285,11 @@ impl CompilerKind {
             | Language::CxxHeader
             | Language::GenericHeader
             | Language::ObjectiveC
-            | Language::ObjectiveCxx => "C/C++",
+            | Language::ObjectiveCxx
+            | Language::ObjectiveCxxHeader => "C/C++",
             Language::Cuda => "CUDA",
+            Language::Ptx => "PTX",
+            Language::Cubin => "CUBIN",
             Language::Rust => "Rust",
             Language::Hip => "HIP",
         }
@@ -181,8 +302,10 @@ impl CompilerKind {
             CompilerKind::C(CCompilerKind::Diab) => textual_lang + " [diab]",
             CompilerKind::C(CCompilerKind::Gcc) => textual_lang + " [gcc]",
             CompilerKind::C(CCompilerKind::Msvc) => textual_lang + " [msvc]",
-            CompilerKind::C(CCompilerKind::Nvhpc) => textual_lang + " [nvhpc]",
             CompilerKind::C(CCompilerKind::Nvcc) => textual_lang + " [nvcc]",
+            CompilerKind::C(CCompilerKind::Cicc) => textual_lang + " [cicc]",
+            CompilerKind::C(CCompilerKind::Ptxas) => textual_lang + " [ptxas]",
+            CompilerKind::C(CCompilerKind::Nvhpc) => textual_lang + " [nvhpc]",
             CompilerKind::C(CCompilerKind::TaskingVX) => textual_lang + " [taskingvx]",
             CompilerKind::Rust => textual_lang,
         }
@@ -274,7 +397,7 @@ where
         rewrite_includes_only: bool,
         storage: Arc<dyn Storage>,
         cache_control: CacheControl,
-    ) -> Result<HashResult>;
+    ) -> Result<HashResult<T>>;
 
     /// Return the state of any `--color` option passed to the compiler.
     fn color_mode(&self) -> ColorMode;
@@ -284,6 +407,7 @@ where
     #[allow(clippy::too_many_arguments)]
     async fn get_cached_or_compile(
         self: Box<Self>,
+        service: &server::SccacheService<T>,
         dist_client: Option<Arc<dyn dist::Client>>,
         creator: T,
         storage: Arc<dyn Storage>,
@@ -335,7 +459,9 @@ where
         // If `ForceRecache` is enabled, we won't check the cache.
         let start = Instant::now();
         let cache_status = async {
-            if cache_control == CacheControl::ForceRecache {
+            if cache_control == CacheControl::ForceNoCache {
+                Ok(Cache::None)
+            } else if cache_control == CacheControl::ForceRecache {
                 Ok(Cache::Recache)
             } else {
                 storage.get(&key).await
@@ -395,6 +521,14 @@ where
                 );
                 Ok(CacheLookupResult::Miss(MissType::Normal))
             }
+            (Ok(Ok(Cache::None)), duration) => {
+                debug!(
+                    "[{}]: Cache none in {}",
+                    out_pretty,
+                    fmt_duration_as_secs(&duration)
+                );
+                Ok(CacheLookupResult::Miss(MissType::ForcedNoCache))
+            }
             (Ok(Ok(Cache::Recache)), duration) => {
                 debug!(
                     "[{}]: Cache recache in {}",
@@ -431,6 +565,7 @@ where
                 let start = Instant::now();
 
                 let (cacheable, dist_type, compiler_result) = dist_or_local_compile(
+                    service,
                     dist_client,
                     creator,
                     cwd,
@@ -446,7 +581,22 @@ where
                         out_pretty,
                         fmt_duration_as_secs(&duration_compilation)
                     );
-                    return Ok((CompileResult::CompileFailed, compiler_result));
+                    return Ok((
+                        CompileResult::CompileFailed(dist_type, duration_compilation),
+                        compiler_result,
+                    ));
+                }
+                if miss_type == MissType::ForcedNoCache {
+                    // Do not cache
+                    debug!(
+                        "[{}]: Compiled in {}, but not caching",
+                        out_pretty,
+                        fmt_duration_as_secs(&duration_compilation)
+                    );
+                    return Ok((
+                        CompileResult::NotCached(dist_type, duration_compilation),
+                        compiler_result,
+                    ));
                 }
                 if cacheable != Cacheable::Yes {
                     // Not cacheable
@@ -455,7 +605,10 @@ where
                         out_pretty,
                         fmt_duration_as_secs(&duration_compilation)
                     );
-                    return Ok((CompileResult::NotCacheable, compiler_result));
+                    return Ok((
+                        CompileResult::NotCacheable(dist_type, duration_compilation),
+                        compiler_result,
+                    ));
                 }
                 debug!(
                     "[{}]: Compiled in {}, storing in cache",
@@ -514,34 +667,36 @@ where
 
 #[cfg(not(feature = "dist-client"))]
 async fn dist_or_local_compile<T>(
+    service: &server::SccacheService<T>,
     _dist_client: Option<Arc<dyn dist::Client>>,
     creator: T,
     _cwd: PathBuf,
-    compilation: Box<dyn Compilation>,
+    compilation: Box<dyn Compilation<T>>,
     _weak_toolchain_key: String,
     out_pretty: String,
 ) -> Result<(Cacheable, DistType, process::Output)>
 where
     T: CommandCreatorSync,
 {
-    let mut path_transformer = dist::PathTransformer::default();
+    let mut path_transformer = dist::PathTransformer::new();
     let (compile_cmd, _dist_compile_cmd, cacheable) = compilation
         .generate_compile_commands(&mut path_transformer, true)
         .context("Failed to generate compile commands")?;
 
     debug!("[{}]: Compiling locally", out_pretty);
     compile_cmd
-        .execute(&creator)
+        .execute(&service, &creator)
         .await
         .map(move |o| (cacheable, DistType::NoDist, o))
 }
 
 #[cfg(feature = "dist-client")]
 async fn dist_or_local_compile<T>(
+    service: &server::SccacheService<T>,
     dist_client: Option<Arc<dyn dist::Client>>,
     creator: T,
     cwd: PathBuf,
-    compilation: Box<dyn Compilation>,
+    compilation: Box<dyn Compilation<T>>,
     weak_toolchain_key: String,
     out_pretty: String,
 ) -> Result<(Cacheable, DistType, process::Output)>
@@ -554,17 +709,17 @@ where
         Some(ref client) => client.rewrite_includes_only(),
         _ => false,
     };
-    let mut path_transformer = dist::PathTransformer::default();
+    let mut path_transformer = dist::PathTransformer::new();
     let (compile_cmd, dist_compile_cmd, cacheable) = compilation
         .generate_compile_commands(&mut path_transformer, rewrite_includes_only)
         .context("Failed to generate compile commands")?;
 
-    let dist_client = match dist_client {
+    let dist_client = match dist_compile_cmd.clone().and(dist_client) {
         Some(dc) => dc,
         None => {
             debug!("[{}]: Compiling locally", out_pretty);
             return compile_cmd
-                .execute(&creator)
+                .execute(service, &creator)
                 .await
                 .map(move |o| (cacheable, DistType::NoDist, o));
         }
@@ -573,8 +728,8 @@ where
     debug!("[{}]: Attempting distributed compilation", out_pretty);
     let out_pretty2 = out_pretty.clone();
 
-    let local_executable = compile_cmd.executable.clone();
-    let local_executable2 = compile_cmd.executable.clone();
+    let local_executable = compile_cmd.get_executable();
+    let local_executable2 = compile_cmd.get_executable();
 
     let do_dist_compile = async move {
         let mut dist_compile_cmd =
@@ -659,7 +814,7 @@ where
             dist::RunJobResult::Complete(jc) => jc,
             dist::RunJobResult::JobNotFound => bail!("Job {} not found on server", job_id),
         };
-        info!(
+        debug!(
             "fetched {:?}",
             jc.outputs
                 .iter()
@@ -735,7 +890,7 @@ where
                 );
 
                 compile_cmd
-                    .execute(&creator)
+                    .execute(service, &creator)
                     .await
                     .map(|o| (DistType::Error, o))
             }
@@ -751,14 +906,21 @@ impl<T: CommandCreatorSync> Clone for Box<dyn CompilerHasher<T>> {
 }
 
 /// An interface to a compiler for actually invoking compilation.
-pub trait Compilation: Send {
+pub trait Compilation<T>: Send
+where
+    T: CommandCreatorSync,
+{
     /// Given information about a compiler command, generate a command that can
     /// execute the compiler.
     fn generate_compile_commands(
         &self,
         path_transformer: &mut dist::PathTransformer,
         rewrite_includes_only: bool,
-    ) -> Result<(CompileCommand, Option<dist::CompileCommand>, Cacheable)>;
+    ) -> Result<(
+        Box<dyn CompileCommand<T>>,
+        Option<dist::CompileCommand>,
+        Cacheable,
+    )>;
 
     /// Create a function that will create the inputs used to perform a distributed compilation
     #[cfg(feature = "dist-client")]
@@ -800,11 +962,14 @@ impl OutputsRewriter for NoopOutputsRewriter {
 }
 
 /// Result of generating a hash from a compiler command.
-pub struct HashResult {
+pub struct HashResult<T>
+where
+    T: CommandCreatorSync,
+{
     /// The hash key of the inputs.
     pub key: String,
     /// An object to use for the actual compilation, if necessary.
-    pub compilation: Box<dyn Compilation + 'static>,
+    pub compilation: Box<dyn Compilation<T> + 'static>,
     /// A weak key that may be used to identify the toolchain
     pub weak_toolchain_key: String,
 }
@@ -854,6 +1019,8 @@ pub enum DistType {
 pub enum MissType {
     /// The compilation was not found in the cache, nothing more.
     Normal,
+    /// Do not cache the results of the compilation.
+    ForcedNoCache,
     /// Cache lookup was overridden, recompilation was forced.
     ForcedRecache,
     /// Cache took too long to respond.
@@ -884,10 +1051,12 @@ pub enum CompileResult {
         Duration, // Compilation time
         Pin<Box<dyn Future<Output = Result<CacheWriteInfo>> + Send>>,
     ),
+    /// Not in cache and do not cache the results of the compilation.
+    NotCached(DistType, Duration),
     /// Not in cache, but the compilation result was determined to be not cacheable.
-    NotCacheable,
+    NotCacheable(DistType, Duration),
     /// Not in cache, but compilation failed.
-    CompileFailed,
+    CompileFailed(DistType, Duration),
 }
 
 /// The state of `--color` options passed to a compiler.
@@ -908,8 +1077,15 @@ impl fmt::Debug for CompileResult {
             CompileResult::CacheMiss(ref m, ref dt, ref d, _) => {
                 write!(f, "CompileResult::CacheMiss({:?}, {:?}, {:?}, _)", d, m, dt)
             }
-            CompileResult::NotCacheable => write!(f, "CompileResult::NotCacheable"),
-            CompileResult::CompileFailed => write!(f, "CompileResult::CompileFailed"),
+            CompileResult::NotCached(ref dt, ref d) => {
+                write!(f, "CompileResult::NotCached({:?}, {:?}_", dt, d)
+            }
+            CompileResult::NotCacheable(ref dt, ref d) => {
+                write!(f, "CompileResult::NotCacheable({:?}, {:?}_", dt, d)
+            }
+            CompileResult::CompileFailed(ref dt, ref d) => {
+                write!(f, "CompileResult::CompileFailed({:?}, {:?})", dt, d)
+            }
         }
     }
 }
@@ -923,8 +1099,11 @@ impl PartialEq<CompileResult> for CompileResult {
             (CompileResult::CacheMiss(m, dt, _, _), CompileResult::CacheMiss(n, dt2, _, _)) => {
                 m == n && dt == dt2
             }
-            (&CompileResult::NotCacheable, &CompileResult::NotCacheable) => true,
-            (&CompileResult::CompileFailed, &CompileResult::CompileFailed) => true,
+            (CompileResult::NotCached(dt, _), CompileResult::NotCached(dt2, _)) => dt == dt2,
+            (CompileResult::NotCacheable(dt, _), CompileResult::NotCacheable(dt2, _)) => dt == dt2,
+            (CompileResult::CompileFailed(dt, _), CompileResult::CompileFailed(dt2, _)) => {
+                dt == dt2
+            }
             _ => false,
         }
     }
@@ -942,6 +1121,8 @@ pub enum Cacheable {
 pub enum CacheControl {
     /// Default caching behavior.
     Default,
+    /// Do not cache the results of the compilation.
+    ForceNoCache,
     /// Ignore existing cache entries, force recompilation.
     ForceRecache,
 }
@@ -981,6 +1162,28 @@ fn is_rustc_like<P: AsRef<Path>>(p: P) -> bool {
             .map(|s| s.to_string_lossy().to_lowercase())
             .as_deref(),
         Some("rustc") | Some("clippy-driver")
+    )
+}
+
+/// Returns true if the given path looks like cicc
+fn is_nvidia_cicc<P: AsRef<Path>>(p: P) -> bool {
+    matches!(
+        p.as_ref()
+            .file_stem()
+            .map(|s| s.to_string_lossy().to_lowercase())
+            .as_deref(),
+        Some("cicc")
+    )
+}
+
+/// Returns true if the given path looks like ptxas
+fn is_nvidia_ptxas<P: AsRef<Path>>(p: P) -> bool {
+    matches!(
+        p.as_ref()
+            .file_stem()
+            .map(|s| s.to_string_lossy().to_lowercase())
+            .as_deref(),
+        Some("ptxas")
     )
 }
 
@@ -1045,6 +1248,30 @@ where
 
     let rustc_executable = if let Some(ref rustc_executable) = maybe_rustc_executable {
         rustc_executable
+    } else if is_nvidia_cicc(executable) {
+        debug!("Found cicc");
+        return CCompiler::new(
+            Cicc {
+                // TODO: Use nvcc --version
+                version: Some(String::new()),
+            },
+            executable.to_owned(),
+            &pool,
+        )
+        .await
+        .map(|c| (Box::new(c) as Box<dyn Compiler<T>>, None));
+    } else if is_nvidia_ptxas(executable) {
+        debug!("Found ptxas");
+        return CCompiler::new(
+            Ptxas {
+                // TODO: Use nvcc --version
+                version: Some(String::new()),
+            },
+            executable.to_owned(),
+            &pool,
+        )
+        .await
+        .map(|c| (Box::new(c) as Box<dyn Compiler<T>>, None));
     } else if is_known_c_compiler(executable) {
         let cc = detect_c_compiler(creator, executable, args, env.to_vec(), pool).await;
         return cc.map(|c| (c, None));
@@ -1100,7 +1327,7 @@ where
         child.arg(&rustc_executable);
     }
 
-    child.env_clear().envs(ref_env(env)).args(&["-vV"]);
+    child.env_clear().envs(env.to_vec()).args(&["-vV"]);
 
     let rustc_vv = run_input_output(child, None).await.map(|output| {
         if let Ok(stdout) = String::from_utf8(output.stdout.clone()) {
@@ -1234,10 +1461,13 @@ where
     let test = b"
 #if defined(__NVCC__) && defined(__NVCOMPILER)
 compiler_id=nvcc-nvhpc
+compiler_version=__CUDACC_VER_MAJOR__.__CUDACC_VER_MINOR__.__CUDACC_VER_BUILD__
 #elif defined(__NVCC__) && defined(_MSC_VER)
 compiler_id=nvcc-msvc
+compiler_version=__CUDACC_VER_MAJOR__.__CUDACC_VER_MINOR__.__CUDACC_VER_BUILD__
 #elif defined(__NVCC__)
 compiler_id=nvcc
+compiler_version=__CUDACC_VER_MAJOR__.__CUDACC_VER_MINOR__.__CUDACC_VER_BUILD__
 #elif defined(_MSC_VER) && !defined(__clang__)
 compiler_id=msvc
 #elif defined(_MSC_VER) && defined(_MT)
@@ -1389,10 +1619,17 @@ compiler_version=__VERSION__
                     "nvcc" => NvccHostCompiler::Gcc,
                     &_ => NvccHostCompiler::Gcc,
                 };
+                let host_compiler_version = lines
+                    .next()
+                    // In case the compiler didn't expand the macro.
+                    .filter(|&line| line != "__VERSION__")
+                    .map(str::to_owned);
+
                 return CCompiler::new(
                     Nvcc {
                         host_compiler,
-                        version: version.clone(),
+                        version,
+                        host_compiler_version,
                     },
                     executable,
                     &pool,
@@ -1460,7 +1697,6 @@ mod test {
     use std::io::{Cursor, Write};
     use std::sync::Arc;
     use std::time::Duration;
-    use std::u64;
     use test_case::test_case;
     use tokio::runtime::Runtime;
 
@@ -1875,7 +2111,7 @@ LLVM version: 6.0",
         let runtime = single_threaded_runtime();
         let pool = runtime.handle();
         let output = "compiler_id=clang\ncompiler_version=\"16.0.0\"";
-        let arguments = vec![
+        let arguments = [
             ovec!["-c", "foo.c", "-o", "foo.o", "-DHELLO"],
             ovec!["-c", "foo.c", "-o", "foo.o", "-DHI"],
             ovec!["-c", "foo.c", "-o", "foo.o"],
@@ -1980,6 +2216,8 @@ LLVM version: 6.0",
         // Write a dummy input file so the preprocessor cache mode can work
         std::fs::write(f.tempdir.path().join("foo.c"), "whatever").unwrap();
         let storage = Arc::new(storage);
+        let service = server::SccacheService::mock_with_storage(storage.clone(), pool.clone());
+
         // Pretend to be GCC.
         next_command(
             &creator,
@@ -2028,6 +2266,7 @@ LLVM version: 6.0",
             .block_on(async {
                 hasher
                     .get_cached_or_compile(
+                        &service,
                         None,
                         creator.clone(),
                         storage.clone(),
@@ -2064,6 +2303,7 @@ LLVM version: 6.0",
             .block_on(async {
                 hasher2
                     .get_cached_or_compile(
+                        &service,
                         None,
                         creator,
                         storage,
@@ -2134,11 +2374,17 @@ LLVM version: 6.0",
         const COMPILER_STDERR: &[u8] = b"compiler stderr";
         let obj = f.tempdir.path().join("foo.o");
         // Dist client will do the compilation
-        let dist_client = Some(test_dist::OneshotClient::new(
+        let dist_client = test_dist::OneshotClient::new(
             0,
             COMPILER_STDOUT.to_owned(),
             COMPILER_STDERR.to_owned(),
-        ));
+        );
+        let service = server::SccacheService::mock_with_dist_client(
+            dist_client.clone(),
+            storage.clone(),
+            pool.clone(),
+        );
+
         let cwd = f.tempdir.path();
         let arguments = ovec!["-c", "foo.c", "-o", "foo.o"];
         let hasher = match c.parse_arguments(&arguments, ".".as_ref(), &[]) {
@@ -2150,7 +2396,8 @@ LLVM version: 6.0",
             .block_on(async {
                 hasher
                     .get_cached_or_compile(
-                        dist_client.clone(),
+                        &service,
+                        Some(dist_client.clone()),
                         creator.clone(),
                         storage.clone(),
                         arguments.clone(),
@@ -2186,7 +2433,8 @@ LLVM version: 6.0",
             .block_on(async {
                 hasher2
                     .get_cached_or_compile(
-                        dist_client.clone(),
+                        &service,
+                        Some(dist_client.clone()),
                         creator,
                         storage,
                         arguments,
@@ -2219,6 +2467,8 @@ LLVM version: 6.0",
         let pool = runtime.handle().clone();
         let storage = MockStorage::new(None, preprocessor_cache_mode);
         let storage: Arc<MockStorage> = Arc::new(storage);
+        let service = server::SccacheService::mock_with_storage(storage.clone(), pool.clone());
+
         // Write a dummy input file so the preprocessor cache mode can work
         std::fs::write(f.tempdir.path().join("foo.c"), "whatever").unwrap();
         // Pretend to be GCC.
@@ -2268,6 +2518,7 @@ LLVM version: 6.0",
         storage.next_get(Err(anyhow!("Some Error")));
         let (cached, res) = runtime
             .block_on(hasher.get_cached_or_compile(
+                &service,
                 None,
                 creator,
                 storage,
@@ -2309,6 +2560,7 @@ LLVM version: 6.0",
         let storage_delay = Duration::from_millis(2);
         let storage = MockStorage::new(Some(storage_delay), preprocessor_cache_mode);
         let storage: Arc<MockStorage> = Arc::new(storage);
+        let service = server::SccacheService::mock_with_storage(storage.clone(), pool.clone());
         // Pretend to be GCC.
         next_command(
             &creator,
@@ -2358,6 +2610,7 @@ LLVM version: 6.0",
         storage.next_get(Ok(Cache::Hit(entry)));
         let (cached, _res) = runtime
             .block_on(hasher.get_cached_or_compile(
+                &service,
                 None,
                 creator,
                 storage,
@@ -2396,6 +2649,7 @@ LLVM version: 6.0",
             CacheMode::ReadWrite,
         );
         let storage = Arc::new(storage);
+        let service = server::SccacheService::mock_with_storage(storage.clone(), pool.clone());
         // Write a dummy input file so the preprocessor cache mode can work
         std::fs::write(f.tempdir.path().join("foo.c"), "whatever").unwrap();
         // Pretend to be GCC.
@@ -2450,6 +2704,7 @@ LLVM version: 6.0",
             .block_on(async {
                 hasher
                     .get_cached_or_compile(
+                        &service,
                         None,
                         creator.clone(),
                         storage.clone(),
@@ -2478,6 +2733,7 @@ LLVM version: 6.0",
         fs::remove_file(&obj).unwrap();
         let (cached, res) = hasher2
             .get_cached_or_compile(
+                &service,
                 None,
                 creator,
                 storage,
@@ -2523,6 +2779,8 @@ LLVM version: 6.0",
             CacheMode::ReadWrite,
         );
         let storage = Arc::new(storage);
+        let service = server::SccacheService::mock_with_storage(storage.clone(), pool.clone());
+
         // Pretend to be GCC.  Also inject a fake object file that the subsequent
         // preprocessor failure should remove.
         let obj = f.tempdir.path().join("foo.o");
@@ -2568,6 +2826,7 @@ LLVM version: 6.0",
             .block_on(async {
                 hasher
                     .get_cached_or_compile(
+                        &service,
                         None,
                         creator,
                         storage,
@@ -2667,12 +2926,19 @@ LLVM version: 6.0",
         };
         // All these dist clients will fail, but should still result in successful compiles
         for dist_client in dist_clients {
+            let service = server::SccacheService::mock_with_dist_client(
+                dist_client.clone(),
+                storage.clone(),
+                pool.clone(),
+            );
+
             if obj.is_file() {
                 fs::remove_file(&obj).unwrap();
             }
             let hasher = hasher.clone();
             let (cached, res) = hasher
                 .get_cached_or_compile(
+                    &service,
                     Some(dist_client.clone()),
                     creator.clone(),
                     storage.clone(),
