@@ -37,17 +37,6 @@
 //! assert!(cache.get_mut(&2).is_none());
 //! ```
 //!
-//! The cache can also be limited by an arbitrary metric calculated from its key-value pairs, see
-//! [`LruCache::with_meter`][with_meter] for more information. If the `heapsize` feature is enabled,
-//! this crate provides one such alternate metric&mdash;`HeapSize`. Custom metrics can be written by
-//! implementing the [`Meter`][meter] trait.
-//!
-//! [with_meter]: struct.LruCache.html#method.with_meter
-//! [meter]: trait.Meter.html
-
-#[cfg(feature = "heapsize")]
-extern crate heapsize;
-
 use std::borrow::Borrow;
 use std::collections::hash_map::RandomState;
 use std::fmt;
@@ -152,33 +141,6 @@ impl<K, V> CountableMeterWithMeasure<K, V, ()> for Count {
         None
     }
 }
-
-#[cfg(feature = "heapsize")]
-mod heap_meter {
-    use heapsize::HeapSizeOf;
-    use std::borrow::Borrow;
-
-    /// Size limit based on the heap size of each cache item.
-    ///
-    /// Requires cache entries that implement [`HeapSizeOf`][1].
-    ///
-    /// [1]: https://doc.servo.org/heapsize/trait.HeapSizeOf.html
-    pub struct HeapSize;
-
-    impl<K, V: HeapSizeOf> super::Meter<K, V> for HeapSize {
-        type Measure = usize;
-
-        fn measure<Q: ?Sized>(&self, _: &Q, item: &V) -> usize
-        where
-            K: Borrow<Q>,
-        {
-            item.heap_size_of_children() + ::std::mem::size_of::<V>()
-        }
-    }
-}
-
-#[cfg(feature = "heapsize")]
-pub use heap_meter::*;
 
 /// An LRU cache.
 #[derive(Clone)]
@@ -285,10 +247,9 @@ impl<K: Eq + Hash, V, S: BuildHasher> LruCache<K, V, S, Count> {
     /// assert_eq!(cache.get_mut(&1), None);
     /// assert_eq!(cache.get_mut(&2), Some(&mut "c"));
     /// ```
-    pub fn get_mut<Q: ?Sized>(&mut self, k: &Q) -> Option<&mut V>
+    pub fn get_mut<Q: Hash + Eq + ?Sized>(&mut self, k: &Q) -> Option<&mut V>
     where
         K: Borrow<Q>,
-        Q: Hash + Eq,
     {
         self.map.get_refresh(k)
     }
@@ -367,18 +328,16 @@ impl<K: Eq + Hash, V, S: BuildHasher, M: CountableMeter<K, V>> LruCache<K, V, S,
     /// cache.insert(1, "a");
     /// assert!(cache.contains_key(&1));
     /// ```
-    pub fn contains_key<Q: ?Sized>(&self, key: &Q) -> bool
+    pub fn contains_key<Q: Hash + Eq + ?Sized>(&self, key: &Q) -> bool
     where
         K: Borrow<Q>,
-        Q: Hash + Eq,
     {
         self.map.contains_key(key)
     }
 
-    pub fn get<Q: ?Sized>(&mut self, k: &Q) -> Option<&V>
+    pub fn get<Q: Hash + Eq + ?Sized>(&mut self, k: &Q) -> Option<&V>
     where
         K: Borrow<Q>,
-        Q: Hash + Eq,
     {
         self.map.get_refresh(k).map(|v| v as &V)
     }
@@ -429,10 +388,9 @@ impl<K: Eq + Hash, V, S: BuildHasher, M: CountableMeter<K, V>> LruCache<K, V, S,
     /// assert_eq!(cache.remove(&2), None);
     /// assert_eq!(cache.len(), 0);
     /// ```
-    pub fn remove<Q: ?Sized>(&mut self, k: &Q) -> Option<V>
+    pub fn remove<Q: Hash + Eq + ?Sized>(&mut self, k: &Q) -> Option<V>
     where
         K: Borrow<Q>,
-        Q: Hash + Eq,
     {
         self.map.remove(k).map(|v| {
             self.current_measure = self
@@ -909,20 +867,5 @@ mod tests {
         assert_eq!(cache.size(), 0);
         assert!(!cache.contains_key("foo1"));
         assert!(!cache.contains_key("foo2"));
-    }
-
-    #[cfg(feature = "heapsize")]
-    #[test]
-    fn test_heapsize_cache() {
-        use super::HeapSize;
-
-        let mut cache = LruCache::<&str, (u8, u8, u8), _, _>::with_meter(8, HeapSize);
-        cache.insert("foo1", (1, 2, 3));
-        cache.insert("foo2", (4, 5, 6));
-        cache.insert("foo3", (7, 8, 9));
-        assert!(!cache.contains_key("foo1"));
-        cache.insert("foo2", (10, 11, 12));
-        cache.insert("foo4", (13, 14, 15));
-        assert!(!cache.contains_key("foo3"));
     }
 }
