@@ -1,5 +1,5 @@
 #![deny(rust_2018_idioms)]
-#![cfg(all(feature = "dist-client"))]
+#![cfg(feature = "dist-client")]
 
 use fs_err as fs;
 use std::io::{self, Read, Write};
